CHANGED: BANDWIDTH can now be a float

+++ m3u8/parser.py
@@ -188,14 +188,14 @@ def _parse_stream_inf(line, data, state):
     data['media_sequence'] = None
     atribute_parser = remove_quotes_parser('codecs', 'audio', 'video', 'subtitles')
     atribute_parser["program_id"] = int
-    atribute_parser["bandwidth"] = int
+    atribute_parser["bandwidth"] = float
     atribute_parser["average_bandwidth"] = int
     state['stream_info'] = _parse_attribute_list(protocol.ext_x_stream_inf, line, atribute_parser)
 
 def _parse_i_frame_stream_inf(line, data):
     atribute_parser = remove_quotes_parser('codecs', 'uri')
     atribute_parser["program_id"] = int
-    atribute_parser["bandwidth"] = int
+    atribute_parser["bandwidth"] = float
     iframe_stream_info = _parse_attribute_list(protocol.ext_x_i_frame_stream_inf, line, atribute_parser)
     iframe_playlist = {'uri': iframe_stream_info.pop('uri'),
                        'iframe_stream_info': iframe_stream_info}
